Fix behavior of update command in presence of cluster name wildcards
Use instance name tag everywhere instead of cluster_name.  This might still
create additional instance profiles, per ring, but it's currently not clear
how to avoid that in general.
